fix compiler warnings

+++ cpp/src/plasma/plasma_client.cc
@@ -376,7 +376,7 @@ static inline bool compute_object_hash_parallel(XXH64_state_t *hash_state,
                                                 int64_t nbytes) {
   // Note that this function will likely be faster if the address of data is
   // aligned on a 64-byte boundary.
-  const uint64_t num_threads = kThreadPoolSize;
+  const int num_threads = kThreadPoolSize;
   uint64_t threadhash[num_threads + 1];
   const uint64_t data_address = reinterpret_cast<uint64_t>(data);
   const uint64_t num_blocks = nbytes / BLOCK_SIZE;
